make float constants explicitely float fix some type issues in loops remove redunant ifs on bool types

+++ src/flange.h
@@ -50,8 +50,8 @@ typedef struct tflanger_t {
     unsigned char trails;          // Bypass with trails or not?
 
     //Delay line pointer
-    int dlyWrite;
-    int maxDly;
+    size_t dlyWrite;
+    size_t  maxDly;
 
     //LFO
     float lfoConst;
